LayoutTests/platform/blackberry/fast/dom/Window/slow-unload-handler-expected.txt

frame "theiframe" - has 3 onunload handler(s)
CONSOLE MESSAGE: line 17: ReferenceError: Can't find variable: layoutTestController
This is a test for bug 29193: Need to prevent Javascript sleeps in unload handlers. If successful, PASS should be printed below.

FAIL
